Update the way we load bindings and schemas
dt-bindings and dt-schema are now loaded directly from network.
There is no more devicetree folder in the git repo

Instead, py-dtbindings while trying to find a local copy of bindings, by default in download/bindings folder.
This folder can be modifier as you wish using the path arg of SDTBindings constructor.

Same for dt-schema except that's download location can't be modifier by arg.

A script "clean_bindings.sh" has been added.
Goal is to delete all bindings file that's are not up to date and that leads the lib to fail.
Usually the error came from a dt-schema bad path on a schema that seems to have been moved.